feat(03-01): create test_corr_health.py with health_status tests for all drivers
- Add 6 healthy-state tests: reader, publisher, RPC responder, progress pub, journal pub, connection
- Add 2 unhealthy-state tests: reader closed, publisher closed (retains counts)
- Move health_status tests from test_messenger_reliability.py
- Keep 4 non-health tests in test_messenger_reliability.py for Plan 02
